+++ react-ui-kit/src/Text/Label.test.tsx
@@ -22,7 +22,7 @@ import {THEME_ID} from '../Layout';
 import {matchComponent} from '../test/testUtil';
 import {Label, LabelLink} from './Label';
 
-describe('"Label"', () => {
+describe('"InputLabel"', () => {
   it('renders', () => matchComponent(<Label>Label</Label>));
   it('renders (dark theme)', () => matchComponent(<Label>Label</Label>, THEME_ID.DARK));
 });